wdio-runner: fix mocks
