+++ packages/wdio-logger/tests/.eslintrc
@@ -1,0 +1,5 @@
+{
+  "env": {
+    "jest": true
+  }
+}
