--- packages/wdio-runner/tests/__mocks__/wdio-logger.js
@@ -1,5 +1,0 @@
-export default () => logMock
-export const logMock = {
-    error: jest.fn(),
-    debug: jest.fn()
-}
